+++ rbx_xml/src/types/enumeration.rs
@@ -1,6 +1,6 @@
 use std::io::{Read, Write};
 
-use rbx_dom_weak::types::EnumValue;
+use rbx_dom_weak::types::Enum;
 
 use crate::{
     core::XmlType,
@@ -9,7 +9,7 @@ use crate::{
     serializer_core::XmlEventWriter,
 };
 
-impl XmlType for EnumValue {
+impl XmlType for Enum {
     const XML_TAG_NAME: &'static str = "token";
 
     fn write_xml<W: Write>(&self, writer: &mut XmlEventWriter<W>) -> Result<(), EncodeError> {
@@ -22,7 +22,7 @@ impl XmlType for EnumValue {
             .parse()
             .map_err(|e| reader.error(e))?;
 
-        Ok(EnumValue::from_u32(value))
+        Ok(Enum::from_u32(value))
     }
 }
 
@@ -34,6 +34,6 @@ mod test {
 
     #[test]
     fn round_trip() {
-        test_util::test_xml_round_trip(&EnumValue::from_u32(4654321));
+        test_util::test_xml_round_trip(&Enum::from_u32(4654321));
     }
 }